adicionando lista com dto

+++ src/main/java/com/rodrigo/br/gerenciador/servico/TarefaServico.java
@@ -1,5 +1,8 @@
 package com.rodrigo.br.gerenciador.servico;
 
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -8,6 +11,8 @@ import org.springframework.stereotype.Service;
 import com.rodrigo.br.gerenciador.modelo.Resposta;
 import com.rodrigo.br.gerenciador.modelo.Tarefa;
 import com.rodrigo.br.gerenciador.repository.TarefaRepository;
+
+
 
 @Service
 public class TarefaServico {
@@ -19,8 +24,13 @@ public class TarefaServico {
     private Resposta resposta;
 
     //Método para listar tarefas
-    public Iterable<Tarefa> listar() {
+    /*public Iterable<Tarefa> listar() {
         return tarefaRepository.findAll();
+    }*/
+
+    public List<Tarefa> findAll() {
+        return tarefaRepository.findAll();
+        
     }
 
     //Método para cadastrar tarefas
@@ -46,8 +56,15 @@ public class TarefaServico {
     
     //Método para remover tarefas
     public ResponseEntity<Resposta> remover(int id) {
-        tarefaRepository.deleteById(id);
-        resposta.setMensagem("A Tarefa foi removida com sucesso!");
-        return new ResponseEntity<Resposta>(resposta, HttpStatus.OK);
+        Optional<Tarefa> optional = tarefaRepository.findById(id);
+        if (optional.isPresent()){
+            tarefaRepository.deleteById(id);
+            resposta.setMensagem("A tarefa foi removida com sucesso!");
+            return new ResponseEntity<Resposta>(resposta, HttpStatus.OK);
+
+        }else{
+            resposta.setMensagem("Não existe tarefa cadastrada com esse id!");
+            return new ResponseEntity<Resposta>(resposta, HttpStatus.BAD_REQUEST);
+        }
     }
 }
